// src/app/interfaces/shipping-app/commodity.ts
export interface ICommodity {
    name: string;
    description: string;
    countryOfManufacture: string;
    quantity: number;
    quantityUnits: string;
    qtyUnitLabel: string;
    totalWeight: number;
    totalWeightUnit: string;
    totalWeightUnitLabel: string;
    totalCustomsValue: number;
    unitPrice: string;
    hsCode: string;
}