// src/app/types/enum/session-items.enum.ts
export enum SessionItems {
    MOBILENUMBER = 'mobileNumber',
    ISLOGGEDIN = 'isLoggedIn',
    ISFROMSUMMARY = 'isFromSummary',
    NEWSHIPMENT = 'newShipment',
    SELECTEDCOUNTRY = 'selectedCountry',
    SELECTEDLANGUAGE = 'selectedLanguage',
    DIALINGPREFIX = 'dialingPrefix',
    PREVIOUSURL = 'previousUrl',
    TERMSANDCONDITIONACCEPTED = 'termsConditionAccepted',
    MESSAGECENTRE = 'messageCentre',
    RECIPIENTADDRESSBOOK = 'isLoadedFromAddressBook'
}